requirements.txt
# Core Deep Learning Frameworks
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# YOLO
ultralytics>=8.0.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Science & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
